corrected the lint

+++ contracts/migration/Prime.sol
@@ -20,7 +20,7 @@ import "openzeppelin-contracts-sol8/token/ERC20/extensions/ERC20Capped.sol";
  * @dev   PrimeDAO v2 native contract.
  */
 contract Prime is ERC20Capped {
-    uint supply = 100000000000000000000000000;
+    uint256 supply = 100000000000000000000000000;
     constructor() public ERC20("Prime", "D2D") ERC20Capped(supply) {
         ERC20._mint(msg.sender, supply);
     }
